refactor(document): 拆分块树工具到独立模块并优化实现
- 重构 blockTreeTool 到 src/tools/document/block/ 目录
- 新增类型定义模块 src/tools/types/
- 优化块树构建算法，提升性能
- 改进工具描述结构和错误处理
