+++ src/tools/document/block/index.ts
@@ -1,0 +1,1 @@
+export { blockTreeTool } from './tree';
